Update template to set defines for 0 length/capacity signals instead of just writing 0 in the source
